requirements.txt
# Voice2Gov Backend Dependencies – v2.0
# Production-Level Backend

# ─── Core Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2
beanie==1.26.0

# ─── Authentication & Security ───────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ─── Rate Limiting ───────────────────────────────────────────
slowapi==0.1.9

# ─── AI / NLP ────────────────────────────────────────────────
spacy==3.7.4
# Run after install: python -m spacy download en_core_web_sm

# ─── Image Processing ────────────────────────────────────────
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# ─── Speech-to-Text ──────────────────────────────────────────
vosk==0.3.45
pydub==0.25.1

# ─── Realtime / WebSocket ────────────────────────────────────
python-socketio==5.11.2
aiohttp==3.9.5

# ─── Maps / Geo ──────────────────────────────────────────────
geopy==2.4.1
httpx==0.27.0

# ─── File Storage (Cloud) ────────────────────────────────────
cloudinary==1.40.0
# boto3==1.34.84      # Uncomment for AWS S3 support
# aioboto3==12.3.0    # Uncomment for async S3 support

# ─── Utilities ───────────────────────────────────────────────
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── Testing ─────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
mongomock-motor==0.0.21
